kernel: add sysjmp.{c,h}
