+++ Makefile.rules
@@ -97,6 +97,7 @@ SOURCES += src/streams.c
 SOURCES += src/stringobj.c
 SOURCES += src/syntaxtree.c
 SOURCES += src/sysfiles.c
+SOURCES += src/sysjmp.c
 SOURCES += src/sysmem.c
 SOURCES += src/sysstr.c
 SOURCES += src/system.c
